commands rework progress - Wrote EcoDelegatedBukkitCommand.kt - Wrote EcoHandledCommand.kt - Wrote Eco#createSubCommand - Renamed RegistrableCommandBase.java to PluginCommandBase - Moved most of EcoPluginCommand.kt to EcoHandledCommand.kt - Changed Delegate type in PluginCommand from CommandBase to PluginCommandBase
