--- eco-api/src/main/java/com/willfp/eco/core/command/RegistrableCommandBase.java
@@ -1,7 +1,0 @@
-package com.willfp.eco.core.command;
-
-public interface RegistrableCommandBase extends CommandBase {
-    void register();
-
-    void unregister();
-}
