+++ eco-core/core-backend/src/main/kotlin/com/willfp/eco/internal/command/EcoSubCommand.kt
@@ -1,0 +1,11 @@
+package com.willfp.eco.internal.command
+
+import com.willfp.eco.core.EcoPlugin
+
+class EcoSubCommand(
+    plugin: EcoPlugin, name: String,
+    permission: String,
+    playersOnly: Boolean
+) : EcoHandledCommand(plugin, name, permission, playersOnly) {
+
+}
